+++ scripts/deadcode-exports.baseline.mjs
@@ -120,6 +120,11 @@ export const KNIP_UNUSED_EXPORT_BASELINE = [
   "extensions/irc/src/control-chars.ts: isIrcControlChar",
   "extensions/irc/src/monitor.ts: resolveIrcInboundTarget",
   "extensions/irc/src/runtime.ts: clearIrcRuntime",
+  "extensions/linux-node/src/commands.ts: LinuxNodeCommandDeps",
+  "extensions/linux-node/src/commands.ts: listLinuxVideoDevices",
+  "extensions/linux-node/src/commands.ts: MAX_MEDIA_RAW_BYTES",
+  "extensions/linux-node/src/commands.ts: VideoDevice",
+  "extensions/linux-node/src/executables.ts: createCachedExecutableResolver",
   "extensions/llama-cpp/src/embedding-provider.ts: createLlamaCppMemoryEmbeddingProvider",
   "extensions/llama-cpp/src/embedding-provider.ts: DEFAULT_LLAMA_CPP_EMBEDDING_MODEL",
   "extensions/llama-cpp/src/embedding-provider.ts: formatLlamaCppSetupError",
@@ -528,6 +533,8 @@ export const KNIP_UNUSED_EXPORT_BASELINE = [
   "src/cli/cron-cli/register.cron-simple.ts: loadCronJobForShow",
   "src/cli/daemon-cli/response.ts: buildDaemonHintItems",
   "src/cli/gateway-cli/run.ts: testing",
+  "src/cli/nodes-camera.ts: CameraClipTarget",
+  "src/cli/nodes-camera.ts: CameraSnapTarget",
   "src/cli/nodes-camera.ts: writeUrlToFile",
   "src/cli/plugins-install-command.ts: loadConfigForInstall",
   "src/cli/plugins-list-format.ts: formatPluginLine",
