+++ src/node-host/plugin-node-host.test.ts
@@ -5,7 +5,6 @@ import { resetPluginRuntimeStateForTest, setActivePluginRegistry } from "../plug
 import {
   invokeRegisteredNodeHostCommand,
   listRegisteredNodeHostCapsAndCommands,
-  resolvePluginNodeHostCommandError,
 } from "./plugin-node-host.js";
 
 const availabilityContext = { config: {}, env: {} };
@@ -219,16 +218,5 @@ describe("plugin node-host registry", () => {
     await expect(invokeRegisteredNodeHostCommand("terminal.resume.v1", null)).rejects.toThrow(
       "requires duplex transport",
     );
-  });
-
-  it("preserves explicit command error codes and contains unknown failures", () => {
-    expect(resolvePluginNodeHostCommandError(new Error("CAMERA_UNAVAILABLE: no camera"))).toEqual({
-      code: "CAMERA_UNAVAILABLE",
-      message: "CAMERA_UNAVAILABLE: no camera",
-    });
-    expect(resolvePluginNodeHostCommandError(new Error("boom"))).toEqual({
-      code: "INVALID_REQUEST",
-      message: "Error: boom",
-    });
   });
 });
